feat: social backend (users, follows, likes, comments, notifications, bookmarks, reports, feed, search, posts)
- Alembic 009-012: user profile fields (handle/bio/location),
  follows/likes, comments/notifications/bookmarks/reports,
  rating → NUMERIC(2,1) (half-step ratings)
- New routers: users, follows, likes, comments, notifications,
  bookmarks, reports, feed (+ GET /api/reviews/{id}), search, posts
- Notification service triggers on follow/like/comment
- Dish autocomplete endpoint GET /api/dishes/search by restaurant_place_id
- POST /api/posts accepts RestaurantFromPlace (Google Places) +
  optional dish_id; legacy restaurant_name path kept for scripts
- Admin moderation endpoints GET /api/reports, PATCH /api/reports/{id}
- docker-compose api port 8002 to avoid local collisions

+++ app/models/dish.py
@@ -92,7 +92,7 @@ class DishReview(Base):
     date_tasted: Mapped[date] = mapped_column(Date, nullable=False)
     time_tasted: Mapped[time | None] = mapped_column(Time, nullable=True)
     note: Mapped[str] = mapped_column(Text, nullable=False)
-    rating: Mapped[int] = mapped_column(Integer, nullable=False)
+    rating: Mapped[Decimal] = mapped_column(Numeric(2, 1), nullable=False)
     portion_size: Mapped[PortionSize | None] = mapped_column(
         Enum(PortionSize, name="portion_size"), nullable=True
     )
